Drawer refactored
